+++ digiwf-apps/packages/components/digiwf-form-builder/src/components/DwfFormBuilder.vue
@@ -1,14 +1,14 @@
 <template>
   <div
-      class="d-flex h-100">
+      class="d-flex pa-5"
+      style="max-height: 90vh">
     <dwf-form-builder-elements/>
     <v-card
-        elevation="1"
+        elevation="3"
         class="contentWrapper modeler-area">
       <v-card-title class="pt-1">{{ name }}</v-card-title>
       <v-card-subtitle>{{ description }}</v-card-subtitle>
-
-      <v-list class="h-100">
+      <v-list>
         <draggable
             :list="currentValue.allOf"
             class="list-group"
@@ -28,7 +28,6 @@
                 @input="onSectionChanged"
             />
           </template>
-
         </draggable>
       </v-list>
     </v-card>
@@ -145,11 +144,9 @@ export default defineComponent({
 
 .modeler-area {
   max-width: 1000px;
-  min-height: 908px;
-  margin-bottom: 8px;
+  overflow-y: scroll;
   margin-left: 40px;
   flex-grow: 1;
-  height: 100%;
 }
 
 </style>