+++ digiwf-apps/packages/components/digiwf-form-builder/src/components/modals/DwfEditSectionModal.vue
@@ -27,15 +27,16 @@
 
 <script lang="ts">
 import {DwfFormRenderer, Section} from "@muenchen/digiwf-form-renderer";
-import {defineComponent, reactive, ref} from "vue";
+import {defineComponent, nextTick, reactive, ref} from "vue";
 
 export default defineComponent({
   components: {DwfFormRenderer},
   props: ['schema', 'value'],
   emits: ['saved'],
-  setup(props, {emit, refs, root}) {
+  setup(props, {emit}) {
     const valid = ref(false);
     const dialog = ref(false);
+    const form = ref(null);
     let currentValue = reactive<Section>(props.value);
 
     const saved = () => {
@@ -43,10 +44,10 @@ export default defineComponent({
     }
 
     const onSaveForm = () => {
-      (refs.form as HTMLFormElement).validate();
+      (form.value as HTMLFormElement).validate();
       if (valid.value) {
         dialog.value = true;
-        root.$nextTick(() => {
+        nextTick(() => {
           dialog.value = false;
         });
         saved();
@@ -55,7 +56,7 @@ export default defineComponent({
 
     const onCancelForm = () => {
       dialog.value = true;
-      root.$nextTick(() => {
+      nextTick(() => {
         dialog.value = false;
       });
     }
@@ -68,6 +69,7 @@ export default defineComponent({
       currentValue,
       dialog,
       valid,
+      form,
       saved,
       onSaveForm,
       onCancelForm,