+++ digiwf-libs/digiwf-json-serialization/digiwf-json-serialization-core/src/main/java/io/muenchendigital/digiwf/json/validation/JsonSchemaValidator.java
@@ -14,6 +14,7 @@ import org.json.JSONObject;
 
 import java.util.Collection;
 import java.util.List;
+
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -66,7 +67,12 @@ public class JsonSchemaValidator {
             return errors;
         }
 
-        final ValidationErrorInformation validationErrorInformation = new ValidationErrorInformation(validationException.getPointerToViolation(), validationException.getSchemaLocation(), validationException.getViolatedSchema());
+        final ValidationErrorInformation validationErrorInformation = new ValidationErrorInformation(
+                validationException.getPointerToViolation(),
+                validationException.getSchemaLocation(),
+                validationException.getViolatedSchema(),
+                validationException.getMessage()
+        );
         errors.add(validationErrorInformation);
         return errors;
     }
@@ -79,7 +85,6 @@ public class JsonSchemaValidator {
 //        schema.validate(data);
         Validator validator = buildValidator();
         validator.performValidation(schema, data);
-
     }
 
     private Schema createSchema(final JSONObject schemaObject) {