+++ digiwf-libs/digiwf-json-serialization/digiwf-json-serialization-core/src/main/java/io/muenchendigital/digiwf/json/validation/ValidationErrorInformation.java
@@ -16,4 +16,6 @@ public class ValidationErrorInformation {
 
     private Schema violatedSchema;
 
+    private String message;
+
 }