+++ digiwf-libs/digiwf-json-serialization/digiwf-json-serialization-core/src/test/java/io/muenchendigital/digiwf/json/validation/JsonValidatorTest.java
@@ -48,7 +48,7 @@ public class JsonValidatorTest {
             this.validationService.validate(new JSONObject(rawSchema).toMap(), data);
         });
 
-        assertThat(exception.getMessage()).isEqualTo("ValidationErrorInformation(pointer=#/stringProp1, schemaPath=#/properties/stringProp1, violatedSchema={\"type\":\"string\",\"pattern\":\"^.{1,30}$\"})");
+        assertThat(exception.getMessage()).isEqualTo("ValidationErrorInformation(pointer=#/stringProp1, schemaPath=#/properties/stringProp1, violatedSchema={\"type\":\"string\",\"pattern\":\"^.{1,30}$\"}, message=#/stringProp1: string [fdsfsdafsdafadsfsadfsdafdfdsfsdafsdafadsfsadfsdafd] does not match pattern ^.{1,30}$)");
     }
 
     @Test
@@ -75,7 +75,7 @@ public class JsonValidatorTest {
             this.validationService.validate(new JSONObject(rawSchema).toMap(), data);
         });
 
-        assertThat(exception.getMessage()).isEqualTo("ValidationErrorInformation(pointer=#, schemaPath=#, violatedSchema={\"additionalProperties\":false,\"required\":[\"numberProp1\"],\"properties\":{\"numberProp1\":{\"type\":\"number\"},\"stringProp1\":{\"type\":\"string\",\"pattern\":\"^.{1,30}$\"}}})");
+        assertThat(exception.getMessage()).isEqualTo("ValidationErrorInformation(pointer=#, schemaPath=#, violatedSchema={\"additionalProperties\":false,\"required\":[\"numberProp1\"],\"properties\":{\"numberProp1\":{\"type\":\"number\"},\"stringProp1\":{\"type\":\"string\",\"pattern\":\"^.{1,30}$\"}}}, message=#: required key [numberProp1] not found)");
     }
 
     @Test