+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change Log
 
-## [Unreleased] - yyyy-mm-dd
+## [apps Unreleased] - yyyy-mm-dd
+
+### Added
+
+- (digiwf-form-builder): validation of list items
+
+### Fixed
+
+- (digiwf-form-builder): move field to nested objects
+- (digiwf-form-builder): general styling issues
+
+## [core Unreleased] - yyyy-mm-dd
+
+### Added
+
+- (digiwf-libs) add functionality to extract error information in json serialization
+
+### Fixed
+
+## [core 0.11.0] - 2022-08-30
 
 ### Added
 
@@ -8,8 +27,6 @@
     - add start process endpoint
 - (digiwf-connector) extract api to own package
 - (digiwf-process) add module and api for starting processes via events
-
-### Changed
 
 ### Fixed